pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "langchain-qa-system"
version = "0.1.0"
description = "Advanced LangChain Document Q&A System"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.12",
    "chromadb>=0.4.22",
    "streamlit>=1.29.0",
    "pypdf2>=3.0.1",
    "beautifulsoup4>=4.12.2",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.5.2",
    "faiss-cpu>=1.7.4",
    "sentence-transformers>=2.2.2",
    "pydantic-settings>=2.2.1",
    "openai>=1.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.11.0",
    "isort>=5.12.0",
    "mypy>=1.7.0",
    "pylint>=3.0.2",
    "pytest>=7.4.2",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-asyncio>=0.21.1",
    "httpx>=0.25.0",
]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.pylint]
max-line-length = 88
indent-string = '    '

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=term-missing"

[tool.coverage.run]
source = ["src"]
omit = ["**/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
]
